+++ core/stores/mon/collection_inserter_mock.go
